+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * changed the allowed characters in metric names
 * added a clear deprecation policy
+* added unit tests for `Proxy`